+++ lib/sidekiq/tasks/strategies/rules/disable_with_comment.rb
@@ -1,0 +1,19 @@
+module Sidekiq
+  module Tasks
+    module Strategies
+      module Rules
+        class DisableWithComment < EnableWithComment
+          def respected?(task)
+            !super
+          end
+
+          protected
+
+          def magic_comment_regex
+            /sidekiq-tasks:disable/
+          end
+        end
+      end
+    end
+  end
+end