+++ lib/sidekiq/tasks/strategies/rules/enable_with_comment.rb
@@ -3,24 +3,28 @@ module Sidekiq
     module Strategies
       module Rules
         class EnableWithComment < Base
-          MAGIC_COMMENT_REGEX = /sidekiq-tasks:enable/
-
           def respected?(task)
-            file, start_line = task.locations.first.split(":")
-            start_line_counting_desc = start_line.to_i > 2 ? start_line.to_i - 3 : 0
-            lines = File.read(file).split("\n")[start_line_counting_desc..start_line_counting_desc + 1].reverse
+            lines = relevant_lines(task)
 
-            valid_magic_comment_line?(lines)
-          rescue Errno::ENOENT
-            raise ArgumentError, "File '#{file}' not found"
+            return false if lines.first.match?(/namespace/)
+
+            lines.any? { |line| line.strip.match?(magic_comment_regex) }
+          end
+
+          protected
+
+          def magic_comment_regex
+            /sidekiq-tasks:enable/
           end
 
           private
 
-          def valid_magic_comment_line?(lines)
-            return false if lines.first.match?(/namespace/)
-
-            lines.any? { |line| line.strip.match?(MAGIC_COMMENT_REGEX) }
+          def relevant_lines(task)
+            file, start_line = task.locations.first.split(":")
+            start_line_counting_desc = start_line.to_i > 2 ? start_line.to_i - 3 : 0
+            File.read(file).split("\n")[start_line_counting_desc..start_line_counting_desc + 1].reverse
+          rescue Errno::ENOENT
+            raise ArgumentError, "File '#{file}' not found"
           end
         end
       end